@@ -10,13 +10,13 @@
     "semver": "7.3.5",
     "shelljs": "^0.8.4",
     "ts-node": "^9.0.0",
-    "typescript": "~4.1.2"
+    "typescript": "~4.4.2"
   },
   "devDependencies": {
-    "@types/node": "^12.0.0",
+    "@types/node": "^16.11.0",
     "@types/semver": "7.3.5",
     "@types/shelljs": "^0.8.8",
-    "typescript": "~4.1.2",
+    "typescript": "~4.4.2",
     "tslib": "^2.3.0"
   }
 }
