@@ -22,15 +22,14 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.1.10",
-    "@types/classnames": "^2.2.11",
     "@types/jest": "^26.0.22",
-    "@types/node": "^12.0.0",
+    "@types/js-cookie": "^2.2.7",
+    "@types/node": "^16.11.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.3",
     "@types/react-router": "^5.1.13",
     "@types/react-router-dom": "^5.1.7",
-    "@types/uuid": "^3.4.5",
-    "@types/js-cookie": "^2.2.7",
+    "@types/uuid": "^8.3.1",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
     "autoprefixer": "^9.8.6",
@@ -42,7 +41,7 @@
     "tailwind-underline-utils": "^1.1.2",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.4",
     "tailwindcss-filters": "^3.0.0",
-    "typescript": "~4.1.2",
+    "typescript": "~4.4.2",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
