@@ -24,19 +24,19 @@
     "moment": "^2.21.0",
     "mysql": "^2.15.0",
     "progress": "^2.0.1",
-    "reflect-metadata": "^0.1.10"
+    "reflect-metadata": "^0.1.13"
   },
   "devDependencies": {
-    "@types/argparse": "^1.0.35",
-    "@types/mysql": "^2.15.0",
-    "@types/progress": "^2.0.1",
+    "@types/argparse": "^2.0.10",
     "@types/chai": "^4.0.4",
     "@types/mocha": "^5.2.7",
-    "rimraf": "^2.6.1",
+    "@types/mysql": "^2.15.0",
+    "@types/progress": "^2.0.1",
     "chai": "^4.1.2",
-    "ts-node": "<7.0.0",
     "mocha": "^4.1.0",
     "mocha-typescript": "^1.1.17",
-    "typescript": "^3.1.3"
+    "rimraf": "^2.6.1",
+    "ts-node": "<7.0.0",
+    "typescript": "^4.4.2"
   }
 }
