@@ -21,14 +21,14 @@
   },
   "dependencies": {
     "@gitpod/gitpod-protocol": "0.1.5",
-    "amqplib": "^0.5.2",
+    "amqplib": "^0.8.0",
     "opentracing": "^0.14.4",
-    "uuid": "^3.3.3"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@types/amqplib": "^0.5.7",
-    "@types/uuid": "^3.4.5",
+    "@types/amqplib": "^0.8.2",
+    "@types/uuid": "^8.3.1",
     "rimraf": "^2.6.1",
-    "typescript": "~4.1.2"
+    "typescript": "~4.4.2"
   }
 }
