@@ -10,21 +10,22 @@
         "src"
     ],
     "devDependencies": {
-        "@types/analytics-node": "^3.1.4",
-        "@types/chai-subset": "^1.3.1",
+        "@types/analytics-node": "^3.1.5",
+        "@types/chai-subset": "^1.3.3",
         "@types/cookie": "^0.4.1",
         "@types/js-yaml": "^3.10.1",
         "@types/mocha": "^5.2.7",
+        "@types/node": "^16.11.6",
         "@types/random-number-csprng": "^1.0.0",
-        "@types/uuid": "^3.4.5",
+        "@types/uuid": "^8.3.1",
         "@types/ws": "^5.1.2",
-        "chai": "^4.1.2",
+        "chai": "^4.3.4",
         "chai-subset": "^1.6.0",
         "mocha": "^5.0.0",
         "mocha-typescript": "^1.1.11",
         "rimraf": "^2.6.2",
         "ts-node": "^9.0.0",
-        "typescript": "~4.1.2",
+        "typescript": "~4.4.2",
         "typescript-json-schema": "^0.50.1",
         "typescript-parser": "^2.6.1"
     },
@@ -36,20 +37,24 @@
         "watch": "leeway exec --package .:lib --transitive-dependencies --filter-type yarn --components --parallel -- tsc -w --preserveWatchOutput"
     },
     "dependencies": {
+        "@types/react": "17.0.32",
         "ajv": "^6.5.4",
         "analytics-node": "^4.0.1",
         "cookie": "^0.4.1",
-        "inversify": "^5.0.1",
-        "jaeger-client": "3.17.2",
+        "inversify": "^5.1.1",
+        "jaeger-client": "^3.18.1",
         "js-yaml": "^3.10.0",
-        "opentracing": "^0.14.4",
+        "opentracing": "^0.14.5",
         "prom-client": "^13.2.0",
         "random-number-csprng": "^1.0.2",
+        "react": "17.0.2",
+        "react-dom": "17.0.2",
         "reconnecting-websocket": "^4.4.0",
-        "reflect-metadata": "^0.1.10",
-        "uuid": "^3.3.3",
+        "reflect-metadata": "^0.1.13",
+        "uuid": "^8.3.2",
         "vscode-jsonrpc": "^5.0.1",
-        "vscode-languageserver-protocol": "3.15.3",
+        "vscode-languageserver-protocol": "3.16.0",
+        "vscode-languageserver-types": "3.16.0",
         "vscode-uri": "^1.0.1",
         "vscode-ws-jsonrpc": "^0.2.0",
         "ws": "^7.4.6"
