@@ -40,17 +40,17 @@
     "@google-cloud/storage": "^5.6.0",
     "@improbable-eng/grpc-web-node-http-transport": "^0.14.0",
     "@jmondi/oauth2-server": "^1.1.0",
-    "@octokit/rest": "18.5.6",
-    "@probot/get-private-key": "1.1.0",
-    "amqplib": "^0.5.2",
+    "@octokit/rest": "18.6.1",
+    "@probot/get-private-key": "^1.1.1",
+    "amqplib": "^0.8.0",
     "base-64": "^1.0.0",
     "bitbucket": "^2.4.2",
     "body-parser": "^1.18.2",
     "cookie": "^0.4.1",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.4",
-    "deep-equal": "^1.0.1",
-    "deepmerge": "^2.2.1",
+    "deep-equal": "^2.0.5",
+    "deepmerge": "^4.2.2",
     "express": "^4.16.1",
     "express-graphql": "^0.9.0",
     "express-http-proxy": "^1.0.7",
@@ -58,7 +58,7 @@
     "express-session": "^1.15.6",
     "fs-extra": "^10.0.0",
     "gitpod-yml-inferrer": "^1.2.0",
-    "google-protobuf": "3.15.8",
+    "google-protobuf": "^3.18.0-rc.2",
     "graphql": "^14.6.0",
     "graphql-tools": "^4.0.7",
     "heapdump": "^0.3.15",
@@ -75,12 +75,12 @@
     "passport-dummy": "^0.0.1",
     "passport-gitlab2": "5.0.0",
     "passport-http": "^0.3.0",
-    "probot": "11.4.0",
+    "probot": "12.1.1",
     "prom-client": "^13.2.0",
     "rate-limiter-flexible": "^2.2.1",
     "reflect-metadata": "^0.1.10",
     "swot-js": "^1.0.3",
-    "uuid": "^3.1.0",
+    "uuid": "^8.3.2",
     "vscode-ws-jsonrpc": "^0.2.0",
     "ws": "^7.4.6"
   },
@@ -90,45 +90,42 @@
     "@graphql-codegen/typescript": "^1.19.0",
     "@graphql-codegen/typescript-resolvers": "^1.18.0",
     "@testdeck/mocha": "0.1.2",
-    "@types/amqplib": "^0.5.7",
-    "@types/assert": "^0.0.31",
-    "@types/base-64": "^0.1.2",
-    "@types/body-parser": "^1.16.5",
-    "@types/chai": "^4.1.2",
-    "@types/chai-http": "^3.0.4",
+    "@types/amqplib": "^0.8.2",
+    "@types/assert": "^1.5.6",
+    "@types/base-64": "^1.0.0",
+    "@types/body-parser": "^1.19.1",
+    "@types/chai": "^4.2.21",
     "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.3",
+    "@types/cors": "^2.8.12",
     "@types/deep-equal": "^1.0.1",
-    "@types/dotenv": "^4.0.2",
-    "@types/express": "^4.11.1",
-    "@types/express-mysql-session": "^1.2.2",
-    "@types/express-session": "^1.15.8",
-    "@types/fs-extra": "^5.0.4",
+    "@types/express": "^4.17.13",
+    "@types/express-mysql-session": "^2.1.3",
+    "@types/express-session": "1.17.4",
+    "@types/fs-extra": "^9.0.12",
     "@types/heapdump": "^0.3.1",
-    "@types/http-proxy": "^1.17.5",
-    "@types/js-yaml": "^3.10.1",
+    "@types/http-proxy": "^1.17.7",
+    "@types/js-yaml": "^4.0.3",
     "@types/lodash.debounce": "^4.0.6",
-    "@types/minio": "^7.0.2",
+    "@types/minio": "^7.0.10",
     "@types/mocha": "^2.2.45",
-    "@types/node": "^12.18.3",
-    "@types/node-fetch": "^2.5.0",
-    "@types/p-throttle": "^2.1.0",
-    "@types/passport": "^1.0.3",
-    "@types/passport-oauth2": "1.4.7",
-    "@types/uuid": "^3.1.0",
+    "@types/node": "^16.11.0",
+    "@types/node-fetch": "^2.5.12",
+    "@types/passport": "^1.0.7",
+    "@types/passport-oauth2": "^1.4.11",
+    "@types/uuid": "^8.3.1",
     "@types/ws": "^5.1.2",
-    "chai": "^4.1.2",
-    "chai-http": "^4.0.0",
-    "concurrently": "^5.3.0",
+    "chai": "^4.3.4",
+    "chai-http": "^4.3.0",
+    "concurrently": "^6.2.1",
     "expect": "^1.20.2",
     "mocha": "^5.0.0",
     "mocha-typescript": "^1.1.11",
-    "rimraf": "^2.6.1",
+    "rimraf": "^3.0.2",
     "source-map-loader": "^0.2.3",
-    "standard": "^10.0.3",
+    "standard": "^16.0.3",
     "ts-node": "<7.0.0",
-    "tslint": "^5.9.1",
-    "typescript": "~4.1.2",
+    "tslint": "^6.1.3",
+    "typescript": "~4.4.2",
     "watch": "^1.0.2"
   }
 }
