@@ -20,7 +20,7 @@
     "css-loader": "^0.28.11",
     "source-map-loader": "^1.1.0",
     "style-loader": "^0.23.1",
-    "typescript": "~4.1.2",
+    "typescript": "~4.4.2",
     "webpack": "^4.40.0",
     "webpack-cli": "^3.1.1"
   },
