@@ -6,7 +6,6 @@
   "scripts": {
     "start": "node ./dist/src/index.js",
     "start-ee": "node ./dist/ee/src/index.js",
-
     "prepare": "yarn clean && yarn build",
     "build": "npx tsc",
     "build:clean": "yarn clean && yarn build",
@@ -26,21 +25,21 @@
     "@gitpod/gitpod-db": "0.1.5",
     "@gitpod/gitpod-messagebus": "0.1.5",
     "@gitpod/gitpod-protocol": "0.1.5",
+    "@gitpod/ws-daemon": "0.1.5",
     "@gitpod/ws-manager": "0.1.5",
     "@gitpod/ws-manager-bridge-api": "0.1.5",
-    "@gitpod/ws-daemon": "0.1.5",
-    "amqplib": "^0.5.2",
-    "express": "^4.16.1",
+    "amqplib": "^0.8.0",
+    "express": "^4.17.1",
     "inversify": "^5.0.1",
     "prom-client": "^13.2.0",
-    "reflect-metadata": "^0.1.10"
+    "reflect-metadata": "^0.1.13"
   },
   "devDependencies": {
-    "@types/amqplib": "^0.5.7",
-    "@types/express": "^4.11.1",
+    "@types/amqplib": "^0.8.2",
+    "@types/express": "^4.17.13",
     "@types/google-protobuf": "^3.7.4",
     "tslint": "^5.9.1",
-    "typescript": "~4.1.2",
+    "typescript": "~4.4.2",
     "watch": "^1.0.2"
   }
 }
