@@ -4,7 +4,7 @@
  * See License-AGPL.txt in the project root for license information.
  */
 
-import uuidv4 = require("uuid/v4");
+import { v4 as uuidv4 } from 'uuid';
 
 export type EMailState = "scheduledInternal"
     | "scheduledSendgrid";
