@@ -52,7 +52,7 @@ export class Baseline1592203031938 implements MigrationInterface {
         // features repositories
         {
             const count = (await queryRunner.query(`SELECT COUNT(1) AS cnt FROM d_b_repository_white_list`))[0].cnt;
-            if (count < 1) {
+            if (Number.parseInt(count) < 1) {
                 const entries = [
                     { url: 'https://github.com/gitpod-io/go-gin-app.git', description: '**Go** - A simple web app implemented in Go and Gin', priority: 7 },
                     { url: 'https://github.com/gitpod-io/rails_sample_app', description: '**Ruby on Rails** - Tutorial sample application', priority: 6 },
