@@ -25,7 +25,7 @@ export async function indexExists(queryRunner: QueryRunner, tableName: string, i
                 AND table_name = '${tableName}'
                 AND index_name = '${indexName}'`
     );
-    return countResult[0].cnt === 1;
+    return Number.parseInt(countResult[0].cnt) === 1;
 };
 
 export async function columnExists(queryRunner: QueryRunner, tableName: string, columnName: string): Promise<boolean> {
@@ -36,7 +36,7 @@ export async function columnExists(queryRunner: QueryRunner, tableName: string, 
                 AND table_name = '${tableName}'
                 AND column_name = '${columnName}'`
     );
-    return countResult[0].cnt === 1;
+    return Number.parseInt(countResult[0].cnt) === 1;
 };
 
 export async function tableExists(queryRunner: QueryRunner, tableName: string): Promise<boolean> {
@@ -46,6 +46,6 @@ export async function tableExists(queryRunner: QueryRunner, tableName: string): 
             WHERE table_schema = '${database}'
                 AND table_name = '${tableName}'`
     );
-    return countResult[0].cnt === 1;
+    return Number.parseInt(countResult[0].cnt) === 1;
 };
 
