@@ -27,7 +27,7 @@ export class BitbucketFileProvider implements FileProvider {
     public async getLastChangeRevision(repository: Repository, revisionOrBranch: string, user: User, path: string): Promise<string> {
         try {
             const api = await this.apiFactory.create(user);
-            const fileMetaData = (await api.repositories.readSrc({ workspace: repository.owner, repo_slug: repository.name, node: revisionOrBranch, path, format: "meta" })).data;
+            const fileMetaData = (await api.repositories.readSrc({ workspace: repository.owner, repo_slug: repository.name, commit: revisionOrBranch, path, format: "meta" })).data;
             return (fileMetaData as any).commit.hash;
         } catch (err) {
             log.error({ userId: user.id }, err);
@@ -42,7 +42,7 @@ export class BitbucketFileProvider implements FileProvider {
 
         try {
             const api = await this.apiFactory.create(user);
-            const contents = (await api.repositories.readSrc({ workspace: commit.repository.owner, repo_slug: commit.repository.name, node: commit.revision, path })).data;
+            const contents = (await api.repositories.readSrc({ workspace: commit.repository.owner, repo_slug: commit.repository.name, commit: commit.revision, path })).data;
             return contents as string;
         } catch (err) {
             log.error({ userId: user.id }, err);
