@@ -14,7 +14,7 @@ export interface IContextParser {
     normalize?(contextUrl: string): string | undefined
     canHandle(user: User, contextUrl: string): boolean
     handle(ctx: TraceContext, user: User, contextUrl: string): Promise<WorkspaceContext>
-    fetchCommitHistory(ctx: TraceContext, user: User, contextUrl: string, commit: string, maxDepth: number): Promise<string[]>
+    fetchCommitHistory(ctx: TraceContext, user: User, contextUrl: string, commit: string, maxDepth: number): Promise<string[] | undefined>
 }
 export const IContextParser = Symbol("IContextParser")
 
@@ -81,7 +81,7 @@ export abstract class AbstractContextParser implements IContextParser {
      *
      * @returns the linear commit history starting from (but excluding) the given commit, in the same order as `git log`
      */
-    public abstract fetchCommitHistory(ctx: TraceContext, user: User, contextUrl: string, commit: string, maxDepth: number): Promise<string[]>;
+    public abstract fetchCommitHistory(ctx: TraceContext, user: User, contextUrl: string, commit: string, maxDepth: number): Promise<string[] | undefined>;
 }
 
 export interface URLParts {
